changes for review comments.
