+++ features/apimgt/org.wso2.carbon.apimgt.core.feature/src/main/resources/conf_templates/org.wso2.carbon.apimgt.core.default.json
@@ -42,7 +42,7 @@
   "apim.cache.restapi_token.expiry_time" : "300s",
   "apim.cache.restapi_cache_control_headers.max_age" : "86400s",
   "apim.analytics.enable": false,
-  "apim.api_quota_limit.enable": true,
+  "apim.api_quota_limit.enable": false,
   "apim.http_client.max_total": "100",
   "apim.http_client.default_max_per_route": "50",
   "apim.key_manager.service_url": "https://localhost:${mgt.transport.https.port}${carbon.context}services/",
